Support dynamic docgen in Description as well
(plus refactoring)
